send survey question answers
