+++ backend/graph/api/schema/mutations.graphqls
@@ -1,0 +1,28 @@
+type MutationRoot {
+    setDevicePushToken(token: String!, languages: [String!]!): Device
+    setEpisodeProgress(id: ID!, progress: Int, duration: Int, context: EpisodeContext): Episode!
+
+    sendSupportEmail(title: String!, content: String!, html: String!): Boolean!
+
+    completeTask(
+        id: ID!,
+        selectedAlternatives: [String!],
+    ): Boolean!
+
+    lockLessonAnswers(id: ID!) : Boolean!
+
+    sendTaskMessage(taskId: ID!, message: String): ID!
+    updateTaskMessage(id: ID!, message: String!): ID!
+
+    sendEpisodeFeedback(episodeId: ID!, message: String, rating: Int): ID!
+    updateEpisodeFeedback(id: ID!, message: String, rating: Int): ID!
+
+    confirmAchievement(id: ID!): ConfirmAchievementResult!
+
+    answerSurveyQuestion(id: UUID!, answer: String!): AnswerSurveyQuestionResult!
+    updateSurveyQuestionAnswer(key: String!, answer: String!): AnswerSurveyQuestionResult!
+}
+
+type AnswerSurveyQuestionResult {
+    id: String!
+}
