+++ backend/graph/api/schema/schema.graphqls
@@ -129,25 +129,3 @@ type QueryRoot{
 
   surveys: [Survey!]!
 }
-
-type MutationRoot {
-  setDevicePushToken(token: String!, languages: [String!]!): Device
-  setEpisodeProgress(id: ID!, progress: Int, duration: Int, context: EpisodeContext): Episode!
-
-  sendSupportEmail(title: String!, content: String!, html: String!): Boolean!
-
-  completeTask(
-    id: ID!,
-    selectedAlternatives: [String!],
-  ): Boolean!
-
-  lockLessonAnswers(id: ID!) : Boolean!
-
-  sendTaskMessage(taskId: ID!, message: String): ID!
-  updateTaskMessage(id: ID!, message: String!): ID!
-
-  sendEpisodeFeedback(episodeId: ID!, message: String, rating: Int): ID!
-  updateEpisodeFeedback(id: ID!, message: String, rating: Int): ID!
-
-  confirmAchievement(id: ID!): ConfirmAchievementResult!
-}
