+++ backend/sqlc/models.go
@@ -1414,7 +1414,7 @@ type UsersSurveyquestionanswer struct {
 	ID         string         `db:"id" json:"id"`
 	QuestionID uuid.UUID      `db:"question_id" json:"questionID"`
 	Answer     null_v4.String `db:"answer" json:"answer"`
-	UpdatedAt  null_v4.Time   `db:"updated_at" json:"updatedAt"`
+	UpdatedAt  time.Time      `db:"updated_at" json:"updatedAt"`
 }
 
 type UsersTaskanswer struct {
